--- src/Code/RecipeBook.java
@@ -1,5 +1,0 @@
-package Code;
-
-public class RecipeBook {
-
-}
